Add dedicated training popup usage
