+++ Scripts/Menu/Popups/RankedMatchPopup.cs
@@ -1,135 +1,42 @@
-using Dawnshard.Menu;
-using Dawnshard.Network;
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 using Dawnshard.Database;
-using MoreMountains.Feedbacks;
 using TMPro;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 using UnityEngine.UI;
-using Dawnshard.Views;
-using Dawnshard.Presenters;
 
 namespace Dawnshard.Menu
 {
-    public class RankedMatchPopup : Popup
+    public class RankedMatchPopup : MatchPopup
     {
-        [SerializeField] protected Button findMatchButton;
-        [SerializeField] protected MMFeedbacks openAnimation;
-        [SerializeField] protected MMFeedbacks waitingAnimation;
-        [SerializeField] protected Transform closePosition;
-        [SerializeField] protected GameObject deckParent;
-        [SerializeField] protected TMP_Text timerText;
-        [SerializeField] protected TMP_Text titleText;
-        [SerializeField] protected Image rankImage = null;
-        [SerializeField] protected Image rankExp = null;
-        [SerializeField] protected GameObject rankExpBar = null;
-        [SerializeField] protected TMP_Text rankNumber = null;
-        [SerializeField] protected TMP_FontAsset bebasAsset;
+        [SerializeField] private Image rankImage = null;
+        [SerializeField] private Image rankExp = null;
+        [SerializeField] private GameObject rankExpBar = null;
+        [SerializeField] private TMP_Text rankNumber = null;
+        [SerializeField] private TMP_FontAsset bebasAsset;
 
-
-        protected DeckPresenter deckPresenter;
-        private bool isOpen = false;
-        public bool isSearchingForMatch = false;
-        private float timer;
-
-        public bool IsSinglePlayer { get; set; } = false;
         public bool IsReverseMode { get; set; } = false;
 
-        protected Action OnPopupClosed;
-
-        protected override void Start()
+        protected override void StartMatchAsync()
         {
-            timer = 0f;
-            base.Start();
-            findMatchButton.onClick.AddListener(ManageSearch);
+            FindMatchAsync();
         }
 
-        private void ManageSearch()
-        {
-            if (!isSearchingForMatch)
-            {
-                if (IsSinglePlayer)
-                    StartAIMatchAsync();
-                else
-                    FindMatchAsync();
-            }
-            else
-            {
-                CancelMatchAsync();
-            }
-            //findMatchButton.interactable = false;
-            //findMatchButton.interactable = true;
-        }
-
-        private void Update()
-        {
-            if (isSearchingForMatch)
-            {
-                timer += Time.deltaTime;
-                float minutes = Mathf.FloorToInt(timer / 60);
-                float seconds = Mathf.FloorToInt(timer % 60);
-                timerText.text = string.Format("{0:00}:{1:00}", minutes, seconds);
-            }
-            else
-            {
-                timer = 0f;
-            }
-        }
-
-        public virtual void SetDeckView(DeckModel deck, Action OnEnd)
-        {
-            if (deckPresenter == null)
-            {
-                deckPresenter = DeckFactory.Instance.CreateDeckView(deck, deckParent.transform);
-            }
-            else
-            {
-                deckPresenter.Model = deck;
-                deckPresenter.UpdateView();
-            }
-            OnPopupClosed = OnEnd;
-        }
-
-        protected virtual async void StartAIMatchAsync()
-        {
-            PlayWaitingAnimation();
-
-            try
-            {
-                await GameController.Instance.StartAIMatch(deckPresenter.Model.Name, false);
-                isSearchingForMatch = true;
-            }
-            catch (Exception e)
-            {
-                Debug.LogException(e);
-                ShowError(e.Message);
-                return;
-            }
-        }
-
-        public async void FindMatchAsync()
+        private async void FindMatchAsync()
         {
             try
             {
                 findMatchButton.interactable = false;
                 closeButton.interactable = false;
-                timer = 0f;
-                float minutes = Mathf.FloorToInt(timer / 60);
-                float seconds = Mathf.FloorToInt(timer % 60);
-                timerText.text = string.Format("{0:00}:{1:00}", minutes, seconds);
+                timerText.text = "00:00";
 
-                await GameController.Instance.FindMatch(deckPresenter.Model.Name, IsReverseMode ? Constants.ReverseMode : Constants.RankedMode);
+                await GameController.Instance.FindMatch(deckPresenter.Model.Name,
+                    IsReverseMode ? Constants.ReverseMode : Constants.RankedMode);
+
                 waitingAnimation.Events.OnComplete.AddListener(() =>
                 {
                     isSearchingForMatch = true;
                     findMatchButton.interactable = true;
                     closeButton.interactable = true;
-
                     waitingAnimation.Events.OnComplete.RemoveAllListeners();
                 });
                 PlayWaitingAnimation();
@@ -138,11 +45,10 @@ namespace Dawnshard.Menu
             {
                 Debug.LogException(e);
                 ShowError(e.Message);
-                return;
             }
         }
 
-        public async void CancelMatchAsync()
+        protected override async void CancelMatchAsync()
         {
             if (!isSearchingForMatch)
                 return;
@@ -167,22 +73,13 @@ namespace Dawnshard.Menu
 
         public override void Open()
         {
-            if (isOpen)
-                return;
             base.Open();
-            openAnimation.PlayFeedbacks();
-            isOpen = true;
             if (rankExp is null || rankImage is null) return;
             SetRankImage();
-            timer = 0f;
             if (IsReverseMode)
-            {
                 titleText.text = Constants.ReverseMode + " Match";
-            }
             else
-            {
                 titleText.text = Constants.RankedMode + " Match";
-            }
         }
 
         private void SetRankImage()
@@ -226,79 +123,6 @@ namespace Dawnshard.Menu
                 rankExp.fillAmount = 0;
                 rankNumber.gameObject.SetActive(false);
             }
-        }
-
-        public void SetOnCloseAction(Action OnClose)
-        {
-            OnPopupClosed += OnClose;
-        }
-
-        public override void Close()
-        {
-            if (!isOpen)
-                return;
-            if (!IsSinglePlayer)
-                CancelMatchAsync();
-            StartCoroutine(CloseCoroutine());
-        }
-
-        public void CloseInstantly()
-        {
-            if (!isOpen)
-                return;
-            GetComponent<RectTransform>().position = closePosition.GetComponent<RectTransform>().position;
-            isOpen = false;
-            base.Close();
-            OnPopupClosed();
-        }
-
-        private IEnumerator CloseCoroutine()
-        {
-            openAnimation.PlayFeedbacks();
-            yield return new WaitWhile(() => openAnimation.IsPlaying || isSearchingForMatch);
-            isOpen = false;
-            OnPopupClosed?.Invoke();
-            base.Close();
-        }
-
-        private void OnMatchFound()
-        {
-            StopWaitingAnimation();
-
-            var op = SceneManager.LoadSceneAsync(Constants.GameScene);
-            op.allowSceneActivation = false;
-
-            void onComplete()
-            {
-                op.allowSceneActivation = true;
-            }
-
-            PlayStartMatchAnimation(onComplete);
-        }
-
-
-        private void PlayStartMatchAnimation(Action OnCompleted)
-        {
-            OnCompleted?.Invoke();
-        }
-
-        private void UndoWaitingAnimation()
-        {
-            if (!isSearchingForMatch)
-                return;
-            waitingAnimation.PlayFeedbacks();
-        }
-
-        private void StopWaitingAnimation()
-        {
-            waitingAnimation.StopFeedbacks();
-        }
-
-        private void PlayWaitingAnimation()
-        {
-            if (isSearchingForMatch)
-                return;
-            waitingAnimation.PlayFeedbacks();
         }
     }
 }
